+++ packages/theme_tailor/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.0.4
+- Fix name collision when theme class contains `themes` field
 - Fix the bug where the generator copies encoders below the `@tailor` annotation to the generated class
 
 # 1.0.3